rename homepage component
