+++ src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { HomeComponent } from './homepage/home/home.component';
+import { HomeComponent } from './homepage/homepage.component';
 import { RadioComponent } from './homepage-radio/homepage-radio.component';
 
 const routes: Routes = [
